+++ crates/sncast/tests/e2e/tx_status.rs
@@ -9,25 +9,27 @@ const REVERTED_TX_HASH: &str = "0x00ae35dacba17cde62b8ceb12e3b18f4ab6e103fa2d5e3
 
 #[tokio::test]
 async fn test_incorrect_transaction_hash() {
-    let args = vec!["tx-status", "0x1", "--url", URL];
+    let args = vec!["get", "tx-status", "0x1", "--url", URL];
     let snapbox = runner(&args);
     let output = snapbox.assert().success();
 
     assert_stderr_contains(
         output,
         indoc! {r"
-        Command: tx-status
+        Command: get tx-status
         Error: Failed to get transaction status: Transaction with provided hash was not found (does not exist)
         "},
     );
 }
 
 #[tokio::test]
-async fn test_succeeded() {
+async fn test_succeeded_old_command() {
     let args = vec!["tx-status", SUCCEEDED_TX_HASH, "--url", URL];
     let snapbox = runner(&args);
 
     snapbox.assert().success().stdout_eq(indoc! {r"
+        [WARNING] `sncast tx-status` has moved to `sncast get tx-status`. `sncast tx-status` will be removed in the future.
+
         Success: Transaction status retrieved
         
         Finality Status:  Accepted on L1
@@ -36,8 +38,21 @@ async fn test_succeeded() {
 }
 
 #[tokio::test]
+async fn test_succeeded() {
+    let args = vec!["get", "tx-status", SUCCEEDED_TX_HASH, "--url", URL];
+    let snapbox = runner(&args);
+
+    snapbox.assert().success().stdout_eq(indoc! {r"
+        Success: Transaction status retrieved
+
+        Finality Status:  Accepted on L1
+        Execution Status: Succeeded
+    "});
+}
+
+#[tokio::test]
 async fn test_reverted() {
-    let args = vec!["tx-status", REVERTED_TX_HASH, "--url", URL];
+    let args = vec!["get", "tx-status", REVERTED_TX_HASH, "--url", URL];
     let snapbox = runner(&args);
 
     snapbox.assert().success().stdout_eq(indoc! {r"